Bugfix in the numa aware adj bench

+++ bench/adjacency_lsit/run_adj_list_bench_metall_numa.cpp
@@ -63,7 +63,7 @@ int main(int argc, char *argv[]) {
     std::cout << "sync_time (s)\t" << elapsed_time << std::endl;
 
     std::cout << "Writing profile" << std::endl;
-    for (int i = 0; i < option.segment_file_name_list.size(); ++i) {
+    for (int i = 0; i < (int)option.segment_file_name_list.size(); ++i) {
       managers[i]->profile("/tmp/metall_profile-" + std::to_string(i) + ".log");
     }
   }
